refactor: move ListView code to the correct page
